# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prashiskshan API Documentation",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for Prashiskshan - NEP 2020 compliant internship management platform. This API supports student internship management, company operations, mentor oversight, and administrative functions.",
    "contact": {
      "name": "Prashiskshan Support",
      "email": "support@prashiskshan.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.prashiskshan.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID Token obtained from authentication"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/interviews/start": {
      "post": {
        "summary": "Start interview practice",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/logbooks": {
      "post": {
        "summary": "Submit a new logbook entry",
        "tags": [
          "Logbooks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "internshipId",
                  "weekNumber",
                  "activities",
                  "startDate",
                  "endDate",
                  "hoursWorked"
                ],
                "properties": {
                  "internshipId": {
                    "type": "string"
                  },
                  "weekNumber": {
                    "type": "number"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "hoursWorked": {
                    "type": "number"
                  },
                  "activities": {
                    "type": "string"
                  },
                  "tasksCompleted": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "skillsUsed": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "challenges": {
                    "type": "string"
                  },
                  "learnings": {
                    "type": "string"
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Logbook submitted successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "409": {
            "description": "Logbook already exists for this week"
          }
        }
      }
    },
...
    "/api/mentor/students/{studentId}/applications": {
      "get": {
        "summary": "Get student applications",
        "tags": [
          "Mentors",
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ]
      }
    },
...
    "/api/students/dashboard": {
      "get": {
        "summary": "Get student dashboard",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieve dashboard data including stats, recent activities, and recommendations",
        "responses": {
          "200": {
            "description": "Dashboard data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "student": {
                              "$ref": "#/components/schemas/Student"
                            },
                            "stats": {
                              "type": "object",
                              "properties": {
                                "totalApplications": {
                                  "type": "integer"
                                },
                                "activeInternships": {
                                  "type": "integer"
                                },
                                "completedInternships": {
                                  "type": "integer"
                                },
                                "totalCredits": {
                                  "type": "number"
                                },
                                "totalHours": {
                                  "type": "number"
                                }
                              }
                            },
                            "recentApplications": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Application"
                              }
                            },
                            "recommendedInternships": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Internship"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Students only"
          }
        }
      }
    },
    "/api/students/profile": {
      "get": {
        "summary": "Get profile with credits and history",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/students/internships": {
      "get": {
        "summary": "Browse available internships",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "department",
            "schema": {
              "type": "string"
            },
            "description": "Filter by department"
          },
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "description": "Filter by location"
          },
          {
            "in": "query",
            "name": "workMode",
            "schema": {
              "type": "string",
              "enum": [
                "remote",
                "onsite",
                "hybrid"
              ]
            },
            "description": "Filter by work mode"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "enum": [
                "remote",
                "onsite",
                "hybrid"
              ],
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "minStipend",
            "schema": {
              "type": "number"
            },
            "description": "Minimum stipend"
          },
          {
            "in": "query",
            "name": "maxStipend",
            "schema": {
              "type": "number"
            },
            "description": "Maximum stipend"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          },
          {
            "in": "query",
            "name": "includeMatchScore",
            "schema": {
              "type": "boolean"
            },
            "description": "Include AI match scores"
          }
        ],
        "description": "Get a list of available internships with filtering and pagination",
        "responses": {
          "200": {
            "description": "Internships retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Internship"
                              }
                            },
                            "pagination": {
                              "$ref": "#/components/schemas/Pagination"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/students/internships/recommended": {
      "get": {
        "summary": "Get recommended internships",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Get personalized internship recommendations based on student profile",
        "responses": {
          "200": {
            "description": "Recommendations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "internships": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Internship"
                              }
                            },
                            "reasons": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              },
                              "description": "Reasons why these internships are recommended"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/students/internships/completed": {
      "get": {
        "summary": "Get completed internships with credit request availability",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/students/internships/{internshipId}": {
      "get": {
        "summary": "Get internship details with match analysis",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "internshipId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Internship ID"
          }
        ]
      }
    },
    "/api/students/internships/{internshipId}/apply": {
      "post": {
        "summary": "Apply to internship",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "internshipId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Internship ID"
          }
        ]
      }
    },
    "/api/students/applications": {
      "get": {
        "summary": "Get my applications",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "shortlisted",
                "withdrawn"
              ]
            },
            "description": "Filter by application status"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "description": "Retrieve all applications submitted by the student",
        "responses": {
          "200": {
            "description": "Applications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "applications": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Application"
                              }
                            },
                            "pagination": {
                              "$ref": "#/components/schemas/Pagination"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Apply to internship",
        "description": "Submit an application for an internship",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "internshipId",
                  "coverLetter"
                ],
                "properties": {
                  "internshipId": {
                    "type": "string",
                    "example": "INT-2024001"
                  },
                  "coverLetter": {
                    "type": "string",
                    "example": "I am writing to express my interest..."
                  },
                  "resumeUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://storage.example.com/resumes/resume.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "application": {
                              "$ref": "#/components/schemas/Application"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error or internship not found"
          },
          "409": {
            "description": "Already applied to this internship"
          }
        }
      }
    },
    "/api/students/applications/{applicationId}": {
      "get": {
        "summary": "Get application details",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Application ID"
          }
        ]
      },
      "delete": {
        "summary": "Withdraw application",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Withdraw a submitted application",
        "parameters": [
          {
            "in": "path",
            "name": "applicationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "APP-2024001"
          }
        ],
        "responses": {
          "200": {
            "description": "Application withdrawn successfully"
          },
          "400": {
            "description": "Cannot withdraw application in current status"
          },
          "404": {
            "description": "Application not found"
          }
        }
      }
    },
    "/api/students/modules/recommended": {
      "get": {
        "summary": "Get recommended modules",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Get learning modules recommended for the student",
        "responses": {
          "200": {
            "description": "Recommended modules retrieved successfully"
          }
        }
      }
    },
    "/api/students/modules/start": {
      "post": {
        "summary": "Start a learning module",
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Begin working on a learning module",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "moduleCode"
                ],
                "properties": {
                  "moduleCode": {
                    "type": "string",
                    "example": "CS301"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module started successfully"
          }
        }
      }
    },
...
    "/api/students/:studentId/credit-requests": {
      "post": {
        "summary": "Create a new credit request",
        "tags": [
          "Students",
          "Credit Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get all credit requests for a student",
        "tags": [
          "Students",
          "Credit Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/students/:studentId/credit-requests/:requestId": {
      "get": {
        "summary": "Get credit request details",
        "tags": [
          "Students",
          "Credit Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}